feat: Css modules added to page building components with basic styling
