+++ packages/website/components/blockbuilder/blockbuilder.module.scss
@@ -1,0 +1,3 @@
+.sectional {
+  padding: 6rem 0;
+}
