--- packages/website/components/imageblock.js
@@ -1,21 +1,0 @@
-import clsx from 'clsx';
-
-/**
- * @param {Object} props.block
-*/
-
-export default function ImageBlock({block}) {
-  return (
-    <>
-      { block.src && (
-        <div class="block image-block">
-          <img src={block.src} class="image" />
-          <div
-            class="label">
-            { block.label }
-          </div>
-        </div>
-      )}
-    </>
-  )
-}
