+++ packages/website/components/imageblock/imageblock.module.scss
@@ -1,0 +1,14 @@
+// ///////////////////////////////////////////////////////////////////// General
+.imageBlock {
+  position: relative;
+}
+
+.image {
+  display: block;
+  width: 100%;
+}
+
+.label {
+  text-align: center;
+  margin-top: 2rem;
+}
