+++ packages/website/components/layout.js
@@ -8,6 +8,7 @@ import { getVersion } from 'Lib/api'
 import { getStatusPageSummary } from 'Lib/statuspage-api'
 import { useQuery } from 'react-query'
 
+// ============================================================== Message Banner
 /**
  * @param {any} highlightMessage
  */
@@ -69,6 +70,7 @@ const MessageBanner = ({ highlightMessage }) => {
   return null
 }
 
+// ====================================================================== Layout
 /**
  * @typedef {import('react').ReactChildren} Children
  * @typedef {(props: import('./types.js').LayoutChildrenProps) => Children} ChildrenFn
@@ -92,14 +94,17 @@ export default function Layout({
   data = null,
   highlightMessage,
 }) {
+
   const { isLoggedIn, isLoading, isFetching } = useLoggedIn({
     redirectTo,
     redirectIfFound,
     enabled: needsLoggedIn,
   })
+
   const shouldWaitForLoggedIn = needsLoggedIn && !isLoggedIn
+
   return (
-    <div className={clsx(pageBgColor, 'flex flex-col min-h-screen')}>
+    <div className="master-container">
       <Head>
         <title>{title}</title>
         <meta name="description" content={description} />
@@ -114,6 +119,7 @@ export default function Layout({
         <meta name="twitter:site" content="@protocollabs" />
         <meta name="twitter:creator" content="@protocollabs" />
       </Head>
+
       {shouldWaitForLoggedIn ? (
         <>
           <Navbar isLoggedIn={isLoggedIn} isLoadingUser={isLoading || isFetching} bgColor={navBgColor} />
@@ -133,6 +139,7 @@ export default function Layout({
           <Footer bgColor={footerBgColor} />
         </>
       )}
+
     </div>
   )
 }
