--- packages/website/components/x_hero/index.js
@@ -1,2 +1,0 @@
-import Hero from './x_hero'
-export default Hero
