+++ packages/website/content/pages/index.json
@@ -17,8 +17,10 @@
           {
             "type": "hero",
             "cols": {
-              "num": "col-10"
+              "num": "col-6",
+              "push_left": "off-3"
             },
+            "format": "header",
             "heading": "The easiest way to store data on the decentralized web.",
             "subheading": "Store your data using our simple API. It’s fast, open, and it’s free.",
             "cta": {
