+++ packages/website/pages/index.module.scss
@@ -1,0 +1,3 @@
+.pageIndex {
+  position: relative;
+}
