+++ packages/website/styles/variables.scss
@@ -26,6 +26,7 @@ $gray900: #212121;
 
 $seaShell: #F1F1F1; // pale matte gray
 $steelGray: #1C1E29; // almost black
+$ebony: #03040A; // text black
 
 // ----------------------------------------------------------------- Breakpoints
 $breakpoint_Tiny: 25.9375rem; // 415px
